Replaced Vagrant reload with halt/up.
A VM can be in three states: destroyed, down, up.

Up creates the VM and starts it, or just starts it, if the VM
has already been created.  Unfortunately, an up will not
rerun the Vagrant file, if the VM is already up.

A reload will start a VM if it is stopped or shut it down first
and then start it, thereby rereading the Vagrant file.  Unfortunately,
reload does not create a VM if it has been destroyed.

To ensure there is a VM running at the end of running the Vagrantfile
regardless of the current state of the VM, one must halt and then
up the Vagrant file.