Fixed pt3 script and added pt4 script
pt3 had an duplicate line for the shared folder, so one was removed.

pt4 enables bonding using eth1 and eth2, leaving eth0 for use
with Vagrant's NAT.  eth5 is configured as a host only network, which
is useful as an alternate means of connecting to the VM if one
might desire to disable eth0.